init run

+++ ControlStack/src/Controller.cpp
@@ -329,14 +329,14 @@ int main() {
 //	TX_torques[22] = opt.full_ref(1);
 
 
-	TX_torques[13] = sol(floor(4./4*(mpc_p.N-1))*20);
-	TX_torques[14] = sol(floor(4./4*(mpc_p.N-1))*20+1);
-	TX_torques[15] = sol(floor(3./4*(mpc_p.N-1))*20);
-	TX_torques[16] = sol(floor(3./4*(mpc_p.N-1))*20+1);
-	TX_torques[17] = sol(floor(2./4*(mpc_p.N-1))*20);
-	TX_torques[18] = sol(floor(2./4*(mpc_p.N-1))*20+1);
-	TX_torques[19] = sol(floor(1./4*(mpc_p.N-1))*20);
-	TX_torques[20] = sol(floor(1./4*(mpc_p.N-1))*20+1);
+	TX_torques[13] = sol(int(floor(4./4*(mpc_p.N-1))*20));
+	TX_torques[14] = sol(int(floor(4./4*(mpc_p.N-1))*20+1));
+	TX_torques[15] = sol(int(floor(3./4*(mpc_p.N-1))*20));
+	TX_torques[16] = sol(int(floor(3./4*(mpc_p.N-1))*20+1));
+	TX_torques[17] = sol(int(floor(2./4*(mpc_p.N-1))*20));
+	TX_torques[18] = sol(int(floor(2./4*(mpc_p.N-1))*20+1));
+	TX_torques[19] = sol(int(floor(1./4*(mpc_p.N-1))*20));
+	TX_torques[20] = sol(int(floor(1./4*(mpc_p.N-1))*20+1));
 	TX_torques[21] = opt.full_ref(0);
 	TX_torques[22] = opt.full_ref(1);
